worker/requirements.txt
moviepy==1.0.3
requests
boto3
python-dotenv
numpy<2.0.0
Pillow
scipy
